preprocessing/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
segment-anything>=1.0
numpy>=1.21.0